finish authorization

+++ contacts-backend/controllers/contactController.js
@@ -5,7 +5,7 @@ const Contact = require("./../models/contactModal");
 // access public
 
 const getContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -16,7 +16,7 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("All field are mandatory");
   }
-  const contact = await Contact.create(req.body);
+  const contact = await Contact.create({ ...req.body, user_id: req.user.id });
   res.status(201).json(contact);
 });
 
@@ -35,6 +35,12 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -48,6 +54,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(400);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to delete this");
   }
   const deletedContact = await Contact.findByIdAndDelete(req.params.id);
   if (!deletedContact) {
